Basic EnemyMovement
Implemented a basic enemyMovement that follows the player with the help of a meshNav and meshNavAgent. The agent can evade(run around) basic obstacles but is not able to jump over them, I will add that later if needed. In theory it is pretty cool and stuff, but the problem is that we have to set the whole scene,
(or platform of the enemy) into a meshNav in order to make it work. This enables cool stuff (like said evading, running up stairs and ramps) but could also lead to some heavy extra work (I dunno, till now it worked fine, but we have to see how it works when we add more levels )