Camera Up and Down
Tried various things and implemeted a method that changes the distanceUp var in the camera script, therefore changing the angle of the camera uo/down and keeping the angle of the playerCharacter. TODO as of yet, the camera can limetless increase and decrease in hight, this has to be fixed. I tried to implement some boundaries but while working somehow, they also cause some bugs, so I commented them out.